Fixed distribution of grid squares.
To equally share the width of each individual square inside the container,
the width of the container is divided by the number of squares specified
by the user. This was achieved by using backtick quotes on the style
method, which made it possible to divide the container width by the
number variable.